requirements.txt
paho-mqtt
pyyaml
jc
